+++ src/app/services/theme.service.ts
@@ -1,8 +1,13 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { constructor } from 'marked';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({ providedIn: 'root' })
 export class ThemeService {
   readonly defaultTheme = 'COSMOS';
+
+  constructor(
+    @Inject(PLATFORM_ID) private platformId: Object) {}
 
   changeTheme(theme: string) {
     localStorage.setItem('theme', theme);
@@ -11,7 +16,9 @@ export class ThemeService {
   }
 
   updateTheme() {
-    const theme = localStorage.getItem('theme') ?? this.defaultTheme;
-    document.querySelector('html').setAttribute('theme', theme);
+    if (isPlatformBrowser(this.platformId)) {
+      const theme = localStorage.getItem('theme') ?? this.defaultTheme;
+      document.querySelector('html').setAttribute('theme', theme);
+    }
   }
 }